+++ boom_blox/BoomBlox/Sphere.h
@@ -17,7 +17,7 @@ public:
 	virtual void DoRender() const;
 
 	float GetRadius() const;
-	Eigen::MatrixXf getK();
+	Eigen::MatrixXd getK();
 
 private:
 	float m_radius;
